feat: dedicated functions approach
