+++ apps/www/components/SurveyResults/charts/IndustryChart.tsx
@@ -1,23 +1,5 @@
 import { SurveyChart, buildWhereClause } from '../SurveyChart'
 
-// Helper function to convert filters to the format expected by the database function
-function buildFunctionParams(activeFilters: Record<string, string>) {
-  const params: Record<string, any> = {}
-
-  // Convert single values to arrays for the function parameters
-  if (activeFilters.person_age && activeFilters.person_age !== 'unset') {
-    params.person_age_filter = [activeFilters.person_age]
-  }
-  if (activeFilters.location && activeFilters.location !== 'unset') {
-    params.location_filter = [activeFilters.location]
-  }
-  if (activeFilters.money_raised && activeFilters.money_raised !== 'unset') {
-    params.money_raised_filter = [activeFilters.money_raised]
-  }
-
-  return params
-}
-
 function generateIndustrySQL(activeFilters: Record<string, string>) {
   const whereClause = buildWhereClause(activeFilters, ['industry IS NOT NULL'])
 
@@ -48,7 +30,6 @@ export function IndustryChart() {
       targetColumn="industry"
       filterColumns={['person_age', 'location', 'money_raised']}
       functionName="get_industry_stats"
-      functionParams={buildFunctionParams}
       generateSQLQuery={generateIndustrySQL}
     />
   )
