+++ apps/www/components/SurveyResults/charts/LocationChart.tsx
@@ -1,23 +1,5 @@
 import { SurveyChart, buildWhereClause } from '../SurveyChart'
 
-// Helper function to convert filters to the format expected by the database function
-function buildFunctionParams(activeFilters: Record<string, string>) {
-  const params: Record<string, any> = {}
-
-  // Convert single values to arrays for the function parameters
-  if (activeFilters.person_age && activeFilters.person_age !== 'unset') {
-    params.person_age_filter = [activeFilters.person_age]
-  }
-  if (activeFilters.team_size && activeFilters.team_size !== 'unset') {
-    params.team_size_filter = [activeFilters.team_size]
-  }
-  if (activeFilters.money_raised && activeFilters.money_raised !== 'unset') {
-    params.money_raised_filter = [activeFilters.money_raised]
-  }
-
-  return params
-}
-
 function generateLocationSQL(activeFilters: Record<string, string>) {
   const whereClause = buildWhereClause(activeFilters)
 
@@ -36,7 +18,6 @@ export function LocationChart() {
       targetColumn="location"
       filterColumns={['person_age', 'team_size', 'money_raised']}
       functionName="get_location_stats"
-      functionParams={buildFunctionParams}
       generateSQLQuery={generateLocationSQL}
     />
   )
